refactor: change type
